requirements.txt
fastapi
uvicorn
python-multipart
qdrant-client
azure-storage-blob
sentence-transformers
PyPDF2
pydantic>=2.0
python-dotenv
openai
instructor
PyMuPDF